chore(ebpf): clippy --fix sweep, verifier-budget-guarded
Closes references/marshal-parked-2026-05-07.md [EBPF-CLIPPY-119].
13 files touched, -59 net lines (mostly collapsed match-arm boilerplate
and manual_clamp → .clamp() rewrites).

Procedure (per parking-lot recipe):
1. Snapshot baseline: tmp/bpf-baseline-pre-ebpf-clippy.txt
   monad-cpu-ebpf: 69672 instructions (7% of 900K budget).
2. cd ebpf && cargo clippy --fix --workspace --allow-dirty --allow-staged
3. cargo build --release: green.
4. Re-run scripts/bpf-verifier-check.sh: tmp/bpf-post-ebpf-clippy.txt
   monad-cpu-ebpf: 69863 instructions (+191, still 7% — within budget).
5. Diff: bound references 143 → 142 (manual_clamp collapse), all 10
   programs compile + verifier-pass. Gate: PASSED.

Round Table sign-off (per parking-lot recommendation): Architect (BPF
target / verifier impact), Developer (impl review), BlackMage (security
surface — clippy rewrites are semantically equivalent; no attack surface
change). Computermancer not needed (no UPC-related programs touched).

Note: cargo test on the ebpf workspace fails on a separate
no_std/bpfel-unknown-none lang-item conflict that pre-dates this commit
(reproduces at HEAD~1) — not introduced here. Verifier check is the
authoritative gate for BPF programs.

+++ ebpf/canary-ebpf/src/main.rs
@@ -255,7 +255,7 @@ fn try_canary_xdp(ctx: &XdpContext) -> Result<u32, ()> {
 /// Rewrite L2 headers to route to the endpoint for the given version.
 #[inline(always)]
 fn route_to_version(_ctx: &XdpContext, data: usize, data_end: usize, service_id: u32, version: u8) {
-    let key = ((service_id as u32) & 0xFFFF) | ((version as u32) << 16);
+    let key = (service_id & 0xFFFF) | ((version as u32) << 16);
     if let Some(endpoint) = unsafe { VERSION_ENDPOINTS.get(&key) } {
         // endpoint layout: dst_ip[16], dst_port[2], mac_dst[6]
         // Rewrite Ethernet destination MAC (bytes 18..24 of endpoint)
